Fix deadlock in procedure transaction execution by restructuring locking logic
- Prevent double mutex lock in procedure transactions by skipping tx lock in check_active for procedure tx
- Restructure CALL PROCEDURE execution to avoid holding executor lock during procedure runtime
- Make evaluator_bridge public for expression evaluation access
- Remove unused execute_call_procedure_in_transaction method
